Move proxy-related helper functions to a dedicated file
